fix(tv): render public homepage experience on home, not gated experiences query
The TV home showed "Not authorized to resolve Query.experiences" on launch.
Root cause: the home (and SearchBrowse's "Popular" rail) fetched admin's
top-level Query.experiences, which is intentionally gated to editors
(authScopes hasPermission "read:experiences", VIEWER tier). The public TV app
has no token, so admin correctly rejects it. There is no public list-all
query by design.

Align with the mobile/web pattern: resolve the curated homepage Experience via
the PUBLIC watchSetting.homepageExperience query, then render it through the
PUBLIC experienceBySlug query. The home is now a single curated homepage
experience (hero + carousels), not a launcher over every experience.

- Extract a shared ExperienceRenderer from experience/[slug].tsx so the home
  and detail screen render an Experience identically (kills the divergence
  that caused this bug). Home passes a sticky header slot; detail passes none.
- Add public GET_WATCH_SETTING; remove LIST_EXPERIENCES and the now-dead
  HomeHero. Drop SearchBrowse's gated "Popular experiences" rail (it 401'd and
  rendered empty for the public app); Recent + Browse-topics need no query.
- Adversarial-review fixes: subtract the sticky header height in
  scrollToSection so nav-carousel targets clear the pinned header (home-only;
  detail header height is 0, unchanged); keep the header in the
  loading/error/empty states so Search stays reachable.

Verified on the Apple TV 4K simulator: home, search, experience detail, and
/watch video-details all render with no regressions. Typecheck + lint clean;
231 unit tests pass (added a regression guard locking the home to the public
queries).
