+++ apps/tv/src/lib/queries.ts
@@ -1,11 +1,4 @@
 // Fragments here are kept structurally in sync with apps/mobile/src/lib/queries.ts.
-//
-// LIST_EXPERIENCES specifically has DIVERGED from the mobile copy — TV
-// selects a per-experience VideoHero block for the focus-driven home
-// hero (see the comment on LIST_EXPERIENCES below). Mobile retains
-// the lightweight shape. Re-align when mobile gains the same feature;
-// do NOT copy mobile's LIST_EXPERIENCES back over the TV version
-// without reading this file first.
 
 /**
  * gql.tada typed GraphQL query and fragments for Experience blocks.
@@ -360,48 +353,31 @@ export const GET_WATCH_EXPERIENCE = graphql(
   ],
 )
 
-// ── Listing query (with VideoHero block for focus-driven hero) ────
+// ── Watch setting (public homepage resolution) ──────────────────────
 //
-// LIST_EXPERIENCES powers the TV home screen: both the rail of cards
-// and the top-of-page focus-driven hero. We select the first
-// VideoHeroBlock block per experience so switching the
-// hero on focus requires zero extra round-trips.
+// Mirrors apps/mobile/src/lib/queries.ts GET_WATCH_SETTING. Resolves the
+// configured homepage Experience's slug via the PUBLIC watchSetting query so
+// the TV home can then render it through experienceBySlug (also PUBLIC) — the
+// same pattern mobile and web use for the watch home.
 //
-// Non-VideoHero blocks are still returned over the wire with only
-// __typename, which is cheap (N * blocks * ~30 bytes). For the
-// current experience count (<20), the total payload stays small.
+// This replaced the former LIST_EXPERIENCES home query, which fetched the
+// editor-gated top-level Query.experiences (authScopes hasPermission
+// "read:experiences", VIEWER tier minimum) and so failed with "Not authorized
+// to resolve Query.experiences" for the unauthenticated public TV app. There
+// is no PUBLIC list-all-experiences query by design; the home is a single
+// curated homepage Experience, not a launcher over every experience.
 //
-// If experience count grows past ~30, or payload profiling shows
-// this query exceeding a reasonable size, consider either:
-//   1) Filtering `blocks` server-side to VideoHero only (Strapi
-//      filters on dynamic zones), or
-//   2) Moving to lazy fetch per-focused-card with on-item-focus.
-
-export const LIST_EXPERIENCES = graphql(
-  `
-    query ListExperiences {
-      experiences {
-        id
-        locales {
-          locale
-          documentId: id
-          slug
-          title
-          metaDescription
-          isHomepage
-          ogImageUrl
-          blocks {
-            __typename
-            ... on VideoHeroBlock {
-              ...VideoHeroFields
-            }
-          }
-        }
+// We select only `slug` here; the home loads the full experience via
+// GET_WATCH_EXPERIENCE (PUBLIC, Apollo-cached).
+export const GET_WATCH_SETTING = graphql(`
+  query GetWatchSetting($locale: String!) {
+    watchSetting(locale: $locale) {
+      homepageExperience {
+        slug
       }
     }
-  `,
-  [VideoHeroFragment],
-)
+  }
+`)
 
 // ── Semantic search query ─────────────────────────────────────────
 //
